csi/api: populate ReadAllocs/WriteAllocs fields
The API is missing values for `ReadAllocs` and `WriteAllocs` fields, resulting
in allocation claims not being populated in the web UI. These fields mirror
the fields in `nomad/structs.CSIVolume`, but sending full `Allocation` values
wil bloat the API response, especially if we combine them with the
`Allocations` list of stubs. Returning a separate list of stubs for read and
write would be ideal, but this can't be done without either bloating the API
response with repeated data or causing a panic in previous versions of the
CLI.

The `nomad/structs` fields are persisted with nil values and are
populated during RPC, so we'll mirror that behavior in the HTTP API and
populate the `ReadAllocs` and `WriteAllocs` fields with a map of allocation
IDs, but with null values. The web UI will then create its `ReadAllocations`
and `WriteAllocations` fields by mapping from those IDs to the values in
`Allocations`, instead of flattening the map into a list.